src/humor_classification/requirements.txt
# Data processing
pandas>=2.0.0
numpy>=1.24.0

# LLM & LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
pydantic>=2.0.0

# BERT & PyTorch
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
sentence-transformers>=2.2.0

# Evaluation & Visualization
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0
